requirements.txt
python-dotenv
pandas
openpyxl
xlsxwriter
thefuzz
requests
cryptography
msgraph-sdk
msgraph-core
azure-identity
